Improve screen reader readibility
When using a screen reader to view pages using react-paginate, currently all of
the page number links will only say the page number. It would be nice if the
currently selected page number also said "Page PAGE_NUMBER is your current
page". This can be accomplished using the `aria-label` attribute. It would
further help some users if the consumers of react-paginate could add some extra
text to the end of the `aria-label` to provide additional context to the users.
This context could include things like the page title, or whatever the consumer
wanted.